+++ examples/src/01-basic-rster/basic-rster.ts
@@ -1,0 +1,21 @@
+import rest from "rster";
+import "@rster/worker-express";
+import express from "express";
+
+const api = rest(function () {
+  this.usePing(); // Adds a ping endpoint to our api
+
+  this.get("/hello", function () {
+    this.action(function (req, res) {
+      res.status(200).json({ message: "Hello World!" }).end();
+    });
+  });
+});
+
+const app = express();
+
+app.use(api.express());
+
+app.listen(3000, () => {
+  console.log("Listening on port 3000");
+});