+++ packages/builder/changelog.md
@@ -1,5 +1,25 @@
 # Changelog for @rster/builder
 
+## v0.1.5
+
+- Bugfix: `@rster/builder` throws an error (500) if no body is provided in a request. This is confusing, because the error message is not helpful. This is now fixed.
+
+## v0.1.4
+
+- Bugfix: `@rster/builder` throws an error (500) if no body is provided in a request. This is confusing, because the error message is not helpful. This is now fixed.
+
+## v0.1.3
+
+- Fix `@rster/builder` module type definition
+
+## v0.1.2
+
+- Make `@rster/builder` compatible with `@rster/common@0.1.2`
+
+## v0.1.1
+
+- Fix: Types in `module.ts` and `api.ts` had issues with parametereized method actions.
+
 ## v0.1.0
 
 - Initial release