+++ packages/rster/changelog.md
@@ -1,5 +1,15 @@
 # Changelog for rster
 
+## v0.1.3
+
+- Update bundled dependencies
+  - @rster/basic@0.1.5
+  - @rster/common@0.1.2
+  - @rster/info@0.1.5
+  - @rster/ping@0.1.1
+  - @rster/types@0.1.4
+  - @rster/util@0.1.0
+
 ## v0.1.2
 
 - Export @rster/common